+++ couchbase/src/instance/request.rs
@@ -703,6 +703,29 @@ impl InstanceRequest for AnalyticsRequest {
         unsafe {
             lcb_cmdanalytics_create(&mut command);
             lcb_cmdanalytics_statement(command, statement_encoded.as_ptr(), statement_len);
+            if let Some(options) = self.options {
+                if let Some(timeout) = options.timeout() {
+                    lcb_cmdanalytics_timeout(command, timeout.as_millis() as u32);
+                }
+                if let Some(params) = options.positional_parameters() {
+                    for param in params {
+                        lcb_cmdanalytics_positional_param(command, param.0.as_ptr(), param.1);
+                    }
+                }
+                if let Some(params) = options.named_parameters() {
+                    for param in params {
+                        let key = param.0;
+                        let value = param.1;
+                        lcb_cmdanalytics_named_param(
+                            command,
+                            key.0.as_ptr(),
+                            key.1,
+                            value.0.as_ptr(),
+                            value.1,
+                        );
+                    }
+                }
+            }
             lcb_cmdanalytics_callback(command, Some(analytics_callback));
             lcb_analytics(instance, cookie, command);
             lcb_cmdanalytics_destroy(command);